Notification: separate integration tests into two files
